@@ -1,103 +1,139 @@
 {
+	"bugs": {
+		"url": "https://github.com/HugoMendes98/graph-flow/issues"
+	},
 	"config": {
 		"commitizen": {
 			"path": "./node_modules/cz-conventional-changelog"
 		}
 	},
 	"dependencies": {
-		"@angular/animations": "~16.0.1",
-		"@angular/cdk": "^16.0.1",
-		"@angular/common": "~16.0.1",
-		"@angular/compiler": "~16.0.1",
-		"@angular/core": "~16.0.1",
-		"@angular/forms": "~16.0.1",
-		"@angular/material": "^16.0.1",
-		"@angular/platform-browser": "~16.0.1",
-		"@angular/platform-browser-dynamic": "~16.0.1",
-		"@angular/router": "~16.0.1",
-		"@mikro-orm/cli": "^5.6.8",
-		"@mikro-orm/core": "^5.6.11",
-		"@mikro-orm/migrations": "^5.6.11",
-		"@mikro-orm/nestjs": "^5.1.7",
-		"@mikro-orm/postgresql": "^5.6.11",
-		"@mikro-orm/reflection": "^5.6.16",
-		"@mikro-orm/seeder": "^5.6.16",
-		"@nestjs/common": "9.4.0",
-		"@nestjs/core": "9.4.0",
-		"@nestjs/platform-express": "9.4.0",
-		"@nestjs/swagger": "6.3.0",
-		"@nestjs/terminus": "^9.2.1",
+		"@angular/animations": "16.2.3",
+		"@angular/cdk": "16.2.2",
+		"@angular/common": "16.2.3",
+		"@angular/compiler": "16.2.3",
+		"@angular/core": "16.2.3",
+		"@angular/forms": "16.2.3",
+		"@angular/material": "16.2.2",
+		"@angular/platform-browser": "16.2.3",
+		"@angular/platform-browser-dynamic": "16.2.3",
+		"@angular/router": "16.2.3",
+		"@heap-code/singleton": "^1.1.1",
+		"@mikro-orm/cli": "^5.7.14",
+		"@mikro-orm/core": "^5.7.14",
+		"@mikro-orm/migrations": "^5.7.14",
+		"@mikro-orm/nestjs": "^5.2.0",
+		"@mikro-orm/postgresql": "^5.7.14",
+		"@mikro-orm/reflection": "^5.7.14",
+		"@mikro-orm/seeder": "^5.7.14",
+		"@nestjs/common": "10.1.2",
+		"@nestjs/core": "10.1.2",
+		"@nestjs/jwt": "^10.1.0",
+		"@nestjs/passport": "^10.0.0",
+		"@nestjs/platform-express": "10.1.2",
+		"@nestjs/swagger": "7.1.4",
+		"@nestjs/terminus": "^10.0.1",
 		"@ngx-translate/core": "^15.0.0",
-		"axios": "^1.0.0",
+		"axios": "^1.5.0",
+		"bcryptjs": "^2.4.3",
 		"class-transformer": "^0.5.1",
 		"class-validator": "^0.14.0",
+		"cookie-parser": "^1.4.6",
 		"deepmerge-ts": "^5.1.0",
+		"passport": "^0.6.0",
+		"passport-jwt": "^4.0.1",
+		"passport-local": "^1.0.0",
 		"qs": "^6.11.0",
 		"reflect-metadata": "^0.1.13",
+		"rete": "^2.0.2",
+		"rete-angular-plugin": "^2.0.2",
+		"rete-area-plugin": "^2.0.0",
+		"rete-connection-plugin": "^2.0.0",
+		"rete-readonly-plugin": "^2.0.0",
 		"rxjs": "~7.8.1",
-		"tslib": "^2.3.0",
+		"treeverse": "^3.0.0",
+		"tslib": "^2.6.2",
 		"zone.js": "~0.13.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "~16.0.1",
+		"@4tw/cypress-drag-drop": "^2.2.4",
+		"@angular-devkit/build-angular": "16.2.1",
 		"@angular-eslint/eslint-plugin": "16.0.1",
 		"@angular-eslint/eslint-plugin-template": "16.0.1",
 		"@angular-eslint/template-parser": "16.0.1",
-		"@angular/cli": "~16.0.1",
-		"@angular/compiler-cli": "~16.0.1",
-		"@angular/language-service": "~16.0.1",
-		"@bartholomej/ngx-translate-extract": "^8.0.2",
+		"@angular/cli": "~16.2.0",
+		"@angular/compiler-cli": "16.2.3",
+		"@angular/language-service": "16.2.3",
 		"@compodoc/compodoc": "^1.1.19",
-		"@nestjs/cli": "^9.2.0",
-		"@nestjs/schematics": "9.1.0",
-		"@nestjs/testing": "9.4.0",
-		"@nx/angular": "16.1.4",
-		"@nx/cypress": "16.1.4",
-		"@nx/devkit": "16.1.4",
-		"@nx/eslint-plugin": "16.1.4",
-		"@nx/jest": "16.1.4",
-		"@nx/js": "16.1.4",
-		"@nx/linter": "16.1.4",
-		"@nx/nest": "16.1.4",
-		"@nx/node": "16.1.4",
-		"@nx/plugin": "16.1.4",
-		"@nx/storybook": "16.1.4",
-		"@nx/webpack": "16.1.4",
-		"@nx/workspace": "16.1.4",
-		"@release-it/conventional-changelog": "^5.1.1",
+		"@cspell/eslint-plugin": "^6.31.2",
+		"@darraghor/eslint-plugin-nestjs-typed": "^3.22.6",
+		"@heap-code/concurrency-synchronization": "^1.0.1",
+		"@nestjs/cli": "^10.1.10",
+		"@nestjs/schematics": "10.0.1",
+		"@nestjs/testing": "10.1.2",
+		"@nx/angular": "16.7.4",
+		"@nx/cypress": "16.7.4",
+		"@nx/devkit": "16.7.4",
+		"@nx/eslint-plugin": "16.7.4",
+		"@nx/jest": "16.7.4",
+		"@nx/js": "16.7.4",
+		"@nx/linter": "16.7.4",
+		"@nx/nest": "16.7.4",
+		"@nx/node": "16.7.4",
+		"@nx/plugin": "16.7.4",
+		"@nx/storybook": "16.7.4",
+		"@nx/webpack": "16.7.4",
+		"@nx/workspace": "16.7.4",
+		"@release-it/conventional-changelog": "^7.0.0",
 		"@shopify/eslint-plugin": "^42.1.0",
-		"@storybook/addon-essentials": "7.0.11",
-		"@storybook/angular": "7.0.11",
+		"@storybook/addon-actions": "^7.3.2",
+		"@storybook/addon-essentials": "7.4.0",
+		"@storybook/addon-interactions": "^7.2.1",
+		"@storybook/addon-storysource": "^7.4.0",
+		"@storybook/angular": "7.4.0",
 		"@storybook/core-common": "^7.0.11",
-		"@storybook/core-server": "7.0.11",
+		"@storybook/core-server": "7.4.0",
+		"@storybook/jest": "~0.1.0",
+		"@storybook/test-runner": "^0.13.0",
+		"@storybook/testing-library": "~0.2.0",
 		"@swc-node/register": "^1.6.2",
 		"@swc/core": "^1.3.44",
+		"@testing-library/cypress": "^9.0.0",
+		"@types/bcryptjs": "^2.4.3",
+		"@types/cookie": "^0.5.1",
+		"@types/cookie-parser": "^1.4.3",
 		"@types/dockerode": "^3.3.14",
 		"@types/jest": "29.5.1",
 		"@types/node": "18.13.0",
+		"@types/passport-jwt": "^3.0.9",
+		"@types/passport-local": "^1.0.35",
+		"@types/treeverse": "^3.0.1",
 		"@types/webpack-node-externals": "^3.0.0",
-		"@typescript-eslint/eslint-plugin": "5.59.5",
-		"@typescript-eslint/parser": "5.59.5",
+		"@typescript-eslint/eslint-plugin": "5.62.0",
+		"@typescript-eslint/parser": "5.62.0",
+		"@vendure/ngx-translate-extract": "^8.2.2",
 		"comment-json": "^4.2.3",
-		"cypress": "^12.12.0",
+		"cypress": "^12.17.2",
 		"cz-conventional-changelog": "^3.3.0",
 		"dockerode": "^3.3.4",
-		"eslint": "~8.40.0",
+		"eslint": "8.46.0",
 		"eslint-config-prettier": "8.8.0",
-		"eslint-plugin-cypress": "^2.10.3",
+		"eslint-plugin-cypress": "2.14.0",
 		"eslint-plugin-eslint-comments": "^3.2.0",
 		"eslint-plugin-etc": "^2.0.2",
-		"eslint-plugin-import": "^2.27.5",
-		"eslint-plugin-jest": "^27.2.1",
+		"eslint-plugin-import": "^2.28.1",
+		"eslint-plugin-jest": "^27.2.3",
 		"eslint-plugin-jest-extended": "^2.0.0",
-		"eslint-plugin-jsdoc": "^44.2.3",
+		"eslint-plugin-jsdoc": "^46.5.1",
 		"eslint-plugin-jsonc": "^2.6.0",
 		"eslint-plugin-markdownlint": "^0.4.1",
 		"eslint-plugin-prettier": "^4.2.1",
 		"eslint-plugin-sonarjs": "^0.19.0",
+		"eslint-plugin-sort-attribute-content": "^0.2.1",
+		"eslint-plugin-sort-decorators": "^0.2.3",
 		"eslint-plugin-sort-destructure-keys": "^1.4.0",
 		"eslint-plugin-sort-exports": "^0.8.0",
-		"eslint-plugin-storybook": "^0.6.12",
+		"eslint-plugin-storybook": "^0.6.13",
 		"eslint-plugin-typescript-sort-keys": "^2.1.0",
 		"eslint-plugin-unicorn": "^47.0.0",
 		"eslint-plugin-unused-imports": "^2.0.0",
@@ -107,26 +143,29 @@
 		"jest-environment-jsdom": "29.5.0",
 		"jest-extended": "^3.2.4",
 		"jest-preset-angular": "~13.1.0",
+		"jest-watch-typeahead": "^2.2.2",
 		"jsonc-eslint-parser": "^2.1.0",
-		"nx": "16.1.4",
-		"nx-cloud": "16.0.5",
+		"ngx-translate-testing": "^7.0.0",
+		"nx": "16.7.4",
+		"nx-cloud": "16.3.0",
 		"nx-stylelint": "^15.0.0",
 		"prettier": "^2.6.2",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"release-it": "^15.11.0",
-		"storybook": "^7.0.11",
-		"stylelint": "^15.1.0",
-		"stylelint-config-standard": "^33.0.0",
-		"stylelint-config-standard-scss": "^9.0.0",
+		"release-it": "^16.1.5",
+		"storybook": "^7.4.0",
+		"stylelint": "^15.10.3",
+		"stylelint-config-standard": "^34.0.0",
+		"stylelint-config-standard-scss": "^10.0.0",
 		"stylelint-order": "^6.0.2",
 		"stylelint-prettier": "^3.0.0",
 		"ts-jest": "29.1.0",
 		"ts-node": "10.9.1",
-		"type-fest": "^3.5.7",
-		"typescript": "5.0.4",
-		"webpack": "^5.64.0"
+		"type-fest": "^3.13.1",
+		"typescript": "5.1.6",
+		"webpack": "5.88.2"
 	},
+	"homepage": "https://github.com/HugoMendes98/graph-flow",
 	"license": "MIT",
 	"mikro-orm": {
 		"configPaths": [
