@@ -1,9 +1,17 @@
 import { StorybookConfig } from "@storybook/angular";
 
 const config: StorybookConfig = {
-	addons: ["@storybook/addon-essentials"],
+	addons: [
+		"@storybook/addon-interactions",
+		"@storybook/addon-essentials",
+		{ name: "@storybook/addon-storysource", options: {} }
+	],
 	docs: { autodocs: true },
-	framework: "@storybook/angular",
-	stories: ["../src/app/**/*.stories.mdx", "../src/app/**/*.stories.@(js|ts)"]
+	framework: { name: "@storybook/angular", options: {} },
+	stories: ["../src/app/**/*.stories.@(js|jsx|ts|tsx|mdx)"]
 };
-module.exports = config;
+export default config;
+
+// To customize your webpack configuration you can use the webpackFinal field.
+// Check https://storybook.js.org/docs/react/builders/webpack#extending-storybooks-webpack-config
+// and https://nx.dev/packages/storybook/documents/custom-builder-configs
