@@ -1,8 +1,11 @@
-import axios, { Axios } from "axios";
+import axios, { Axios, AxiosHeaders } from "axios";
 import { Jsonify } from "type-fest";
-import { EntityId } from "~/app/common/dtos/_lib/entity";
-import { FindResultsDto } from "~/app/common/dtos/_lib/find-results.dto";
-import { DropFirst } from "~/app/common/types";
+import { AuthLoginDto } from "~/lib/common/app/auth/dtos";
+import { EntityId } from "~/lib/common/dtos/entity";
+import { FindResultsDto } from "~/lib/common/dtos/find-results.dto";
+import { DropFirst } from "~/lib/common/types";
+
+import { AuthHttpClient } from "../auth.http-client";
 
 /**
  * This is (almost) only a helper for the entrypoint string.
@@ -12,47 +15,69 @@ import { DropFirst } from "~/app/common/types";
  * not to replace the reals API clients.
  */
 export abstract class EntityHttpClient<T> {
+	protected client!: Axios;
+
+	public constructor() {
+		this.removeAuth();
+	}
+
 	/**
 	 * @returns the endpoint
 	 */
 	public abstract getEndpoint(): string;
 
 	public findManyResponse(...params: DropFirst<Parameters<Axios["get"]>>) {
-		return axios.get<Jsonify<FindResultsDto<T>>>(this.getEndpoint(), ...params);
+		return this.client.get<Jsonify<FindResultsDto<T>>>(this.getEndpoint(), ...params);
 	}
 	public findMany(...params: DropFirst<Parameters<Axios["get"]>>) {
 		return this.findManyResponse(...params).then(({ data }) => data);
 	}
 
 	public findOneResponse(id: EntityId, ...params: DropFirst<Parameters<Axios["get"]>>) {
-		return axios.get<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
+		return this.client.get<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
 	}
 	public findOne(id: EntityId, ...params: DropFirst<Parameters<Axios["get"]>>) {
 		return this.findOneResponse(id, ...params).then(({ data }) => data);
 	}
 
 	public createResponse(...params: DropFirst<Parameters<Axios["post"]>>) {
-		return axios.post<Jsonify<T>>(this.getEndpoint(), ...params);
+		return this.client.post<Jsonify<T>>(this.getEndpoint(), ...params);
 	}
 	public create(...params: DropFirst<Parameters<Axios["post"]>>) {
 		return this.createResponse(...params).then(({ data }) => data);
 	}
 
 	public updateResponse(id: EntityId, ...params: DropFirst<Parameters<Axios["patch"]>>) {
-		return axios.patch<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
+		return this.client.patch<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
 	}
-
 	public update(id: EntityId, ...params: DropFirst<Parameters<Axios["patch"]>>) {
 		return this.updateResponse(id, ...params).then(({ data }) => data);
 	}
 
 	public deleteResponse(id: EntityId, ...params: DropFirst<Parameters<Axios["delete"]>>) {
-		return axios.delete<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
+		return this.client.delete<Jsonify<T>>(`${this.getEndpoint()}/${id}`, ...params);
 	}
-
 	public delete(id: EntityId, ...params: DropFirst<Parameters<Axios["delete"]>>) {
 		return this.deleteResponse(id, ...params).then(({ data }) => data);
 	}
 
-	// TODO: auth
+	public async setAuth(email: string, password: string) {
+		const { access_token } = await new AuthHttpClient().login({
+			email,
+			password
+		} satisfies AuthLoginDto);
+
+		this.setAuthToken(access_token);
+		return access_token;
+	}
+
+	public setAuthToken(token: string) {
+		this.client = axios.create({
+			headers: new AxiosHeaders().setAuthorization(`Bearer ${token}`)
+		});
+	}
+
+	public removeAuth() {
+		this.client = axios.create();
+	}
 }
