@@ -1,7 +1,16 @@
 import { JestConfigWithTsJest } from "ts-jest";
 
 export default {
-	coverageDirectory: "../../coverage/apps/backend",
+	collectCoverageFrom: [
+		"<rootDir>/src/**/*.ts",
+		// The controllers are mainly tested with the e2e tests,
+		//	so their code coverage are not really useful
+		"!<rootDir>/src/**/*.controller.ts",
+		"!<rootDir>/src/**/index.ts",
+		// Used to start an application (also in e2e)
+		"!<rootDir>/src/bootstrap.ts",
+		"!<rootDir>/src/main.ts"
+	],
 	displayName: "backend",
 	globalSetup: "<rootDir>/test/support/global/setup.ts",
 	globalTeardown: "<rootDir>/test/support/global/teardown.ts",
