@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { UserCreateDto, UserUpdateDto } from "~/app/common/dtos/user";
+import { UserCreateDto, UserUpdateDto } from "~/lib/common/app/user/dtos";
 
 import { User } from "./user.entity";
 import { UserRepository } from "./user.repository";
@@ -10,7 +10,22 @@ import { EntityService } from "../_lib/entity";
  */
 @Injectable()
 export class UserService extends EntityService<User, UserCreateDto, UserUpdateDto> {
+	/**
+	 * Constructor with "dependency injection"
+	 *
+	 * @param repository injected
+	 */
 	public constructor(repository: UserRepository) {
 		super(repository);
+	}
+
+	/**
+	 * Searches for a user with the given credentials
+	 *
+	 * @param email The email to search
+	 * @returns the user with the given credentials
+	 */
+	public findByCredentials(email: string) {
+		return this.repository.findOneOrFail({ email }, { populate: ["password"] });
 	}
 }
