@@ -1,27 +1,22 @@
-import { EntityManager } from "@mikro-orm/core";
+import { EntityManager, Reference } from "@mikro-orm/core";
 import { Seeder } from "@mikro-orm/seeder";
 import { ReadonlyDeep } from "type-fest";
-import { EntityDto } from "~/app/common/dtos/_lib/entity";
-import { GroupDto } from "~/app/common/dtos/group";
-import { UserDto } from "~/app/common/dtos/user";
+import { EntityDto } from "~/lib/common/dtos/entity";
+import { MockSeed } from "~/lib/common/seeds";
 
 import { EntityBase } from "../../../app/_lib/entity";
-import { Group } from "../../../app/group/group.entity";
+import { AuthService } from "../../../app/auth/auth.service";
+import { Category } from "../../../app/category/category.entity";
+import { GraphArc } from "../../../app/graph/arc/graph-arc.entity";
+import { Graph } from "../../../app/graph/graph.entity";
+import { GraphNode } from "../../../app/graph/node/graph-node.entity";
+import { GraphNodeInput } from "../../../app/graph/node/input";
+import { GraphNodeOutput } from "../../../app/graph/node/output";
+import { NodeInput } from "../../../app/node/input";
+import { Node } from "../../../app/node/node.entity";
+import { NodeOutput } from "../../../app/node/output";
 import { User } from "../../../app/user/user.entity";
-
-/**
- * The values for a DB which all data is mocked
- */
-export interface MockedDb {
-	/**
-	 * Represents the [group]{@link GroupDto} table
-	 */
-	groups: readonly GroupDto[];
-	/**
-	 * Represents the [user]{@link UserDto} table
-	 */
-	users: readonly UserDto[];
-}
+import { Workflow } from "../../../app/workflow/workflow.entity";
 
 /**
  * A seeder that seeds a full DB.
@@ -34,7 +29,7 @@ export abstract class MockedDbSeeder extends Seeder {
 	public static GetMockedDb() {
 		const db = new (this.prototype.constructor as new () => MockedDbSeeder)().db;
 
-		// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- If the function is called
+		// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- Could happen
 		if (!db) {
 			throw new Error(
 				"DB not set! Do not call this function from the `MockedDbSeeder` abstract class."
@@ -47,7 +42,7 @@ export abstract class MockedDbSeeder extends Seeder {
 	/**
 	 * The sample DB to seed
 	 */
-	protected abstract readonly db: ReadonlyDeep<MockedDb>;
+	protected abstract readonly db: ReadonlyDeep<MockSeed>;
 
 	/**
 	 * @inheritDoc
@@ -58,14 +53,44 @@ export abstract class MockedDbSeeder extends Seeder {
 			mocks: readonly EntityDto[];
 		}
 
-		const { groups, users } = this.db;
+		// FIXME: Remove this (with another order or inserts ?)
+		// Disable FK checks
+		await em.getConnection().execute("SET session_replication_role = 'replica';");
+
+		const {
+			categories,
+			graph: { graphArcs, graphNodeInputs, graphNodeOutputs, graphNodes, graphs },
+			node: { nodeInputs, nodeOutputs, nodes },
+			users,
+			workflows
+		} = this.db;
 
 		for (const { entity, mocks } of [
-			{ entity: User, mocks: users },
-			{ entity: Group, mocks: groups }
+			{ entity: Category, mocks: categories },
+			{
+				entity: User,
+				mocks: await Promise.all(
+					users.map(async ({ password, ...user }) => ({
+						...user,
+						password: await AuthService.hash(password)
+					}))
+				)
+			},
+			// Nodes
+			{ entity: Node, mocks: nodes },
+			{ entity: NodeInput, mocks: nodeInputs },
+			{ entity: NodeOutput, mocks: nodeOutputs },
+			// Graphs
+			{ entity: Graph, mocks: graphs },
+			{ entity: GraphNode, mocks: graphNodes },
+			{ entity: GraphNodeInput, mocks: graphNodeInputs },
+			{ entity: GraphNodeOutput, mocks: graphNodeOutputs },
+			{ entity: GraphArc, mocks: graphArcs },
+
+			{ entity: Workflow, mocks: workflows }
 		] satisfies MockEntity[]) {
 			for (const mock of mocks) {
-				em.create<EntityBase>(entity, mock);
+				em.getRepository<EntityBase>(entity).create(mock);
 			}
 
 			// Confirm new rows
@@ -73,7 +98,7 @@ export abstract class MockedDbSeeder extends Seeder {
 
 			// Need to update the sequence when entities are added manually
 			const primaryKey: keyof EntityBase = "_id";
-			// TODO: better (if the table name is set manually)
+			// // TODO: better (if the table name is set manually)
 			const tblName = em.config.getNamingStrategy().classToTableName(entity.name);
 			await em
 				.getConnection()
@@ -84,5 +109,16 @@ export abstract class MockedDbSeeder extends Seeder {
 			// Confirm sequence update
 			await em.flush();
 		}
+
+		// Enable FK checks
+		await em.getConnection().execute("SET session_replication_role = 'origin';");
+
+		for (const { __categories, _id } of nodes) {
+			const node = await em.findOneOrFail(Node, _id, { populate: ["categories"] });
+			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- Does exist with the `populate` option
+			node.categories!.add(__categories.map(id => Reference.createFromPK(Category, id)));
+		}
+
+		// Seeder always flush at the end
 	}
 }
