@@ -1,5 +1,6 @@
+import { BASE_SEED } from "~/lib/common/seeds";
+
 import { MockedDbSeeder } from "./_lib/mocked-db.seeder";
-import { DB_BASE_SEED } from "./seeds";
 
 /**
  * DB Seeder with base sample for seeding.
@@ -8,5 +9,5 @@ export class DbBaseSeeder extends MockedDbSeeder {
 	/**
 	 * @inheritDoc
 	 */
-	protected readonly db = DB_BASE_SEED;
+	protected readonly db = BASE_SEED;
 }
