@@ -1,8 +1,9 @@
 import { JestConfigWithTsJest } from "ts-jest";
 
 export default {
-	coverageDirectory: "../../coverage/apps/frontend",
+	collectCoverageFrom: ["<rootDir>/src/**/*.ts", "!<rootDir>/src/**/index.ts"],
 	displayName: "frontend",
+	moduleNameMapper: { "rete-angular-plugin/16": "rete-angular-plugin" },
 	preset: "../../jest.preset.js",
 	setupFilesAfterEnv: ["<rootDir>/test/setup.ts"],
 	snapshotSerializers: [
