@@ -1,14 +1,44 @@
-import { NgModule } from "@angular/core";
+import { APP_INITIALIZER, ApplicationRef, DoBootstrap, NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { RouterModule } from "@angular/router";
-import { ApiModule } from "~/app/ng/lib/api";
+import { ApiModule } from "~/lib/ng/lib/api";
 
-import { AppComponent } from "./views/_layout/app.component";
-import { ViewsModule } from "./views/views.module";
+import { AppComponent } from "./app.component";
+import { AuthInterceptor } from "./auth/auth.interceptor";
+import { AuthModule } from "./auth/auth.module";
+import { AuthService } from "./auth/auth.service";
+import { environment } from "../environments/environment";
+import { AppRouterModule } from "../lib/router";
+import { AppTranslationModule } from "../lib/translation";
 
 @NgModule({
-	bootstrap: [AppComponent],
-	imports: [ApiModule, BrowserAnimationsModule, BrowserModule, RouterModule, ViewsModule]
+	imports: [
+		AppComponent,
+		AppRouterModule,
+		ApiModule.forRoot({ client: environment.backend }),
+		AppTranslationModule,
+		AuthModule,
+		BrowserAnimationsModule,
+		BrowserModule
+	],
+	providers: [
+		{
+			deps: [AuthInterceptor, AuthService],
+			multi: true,
+			provide: APP_INITIALIZER,
+			useFactory: (interceptor: AuthInterceptor, service: AuthService) => () =>
+				interceptor.runUnprotected(() =>
+					service.refresh().catch((error: unknown) => {
+						if (!AuthService.isAnUnauthorizedError(error)) {
+							throw error;
+						}
+					})
+				)
+		}
+	]
 })
-export class AppModule {}
+export class AppModule implements DoBootstrap {
+	public ngDoBootstrap(appRef: ApplicationRef) {
+		appRef.bootstrap(AppComponent);
+	}
+}
