@@ -1,38 +1,43 @@
 import { APP_INITIALIZER, NgModule, Provider } from "@angular/core";
-import { MatDividerModule } from "@angular/material/divider";
-import { MatExpansionModule } from "@angular/material/expansion";
+import { MatSnackBar } from "@angular/material/snack-bar";
 import { RouterModule } from "@angular/router";
 
-import { StyleguideButtonsComponent } from "./components/styleguide/styleguide-buttons/styleguide-buttons.component";
+import { DevNotifierSnackBar } from "./components/dev-notifier.snack-bar/dev-notifier.snack-bar";
 import { DevView } from "./dev.view";
 import { devRoutes } from "./views/dev.routes";
-import { IndexView } from "./views/index/index.view";
-import { StyleguideView } from "./views/styleguide/styleguide.view";
-import { ComponentsModule } from "../app/components/components.module";
 
 export const DevInitializerProvider: Provider = {
+	deps: [MatSnackBar],
 	multi: true,
 	provide: APP_INITIALIZER,
-	useFactory: () => () => {
-		// TODO: use a snack/dialog/popup/... to notify that this page exists
-		// eslint-disable-next-line no-console -- TODO
-		console.log("!! `_dev` page available !!");
+	useFactory: (snackBar: MatSnackBar) => () => {
+		const value = window.sessionStorage.getItem(DevNotifierSnackBar.SESSION_KEY);
+		if (value === DevNotifierSnackBar.SESSION_NO_MORE) {
+			return;
+		}
+		if (window.location.pathname.startsWith("/_dev")) {
+			// Already in the dev page
+			return;
+		}
+
+		snackBar.openFromComponent(DevNotifierSnackBar, {
+			duration: 10000,
+			horizontalPosition: "center",
+			verticalPosition: "top"
+		});
 	}
 };
 
 @NgModule({
-	declarations: [DevView, IndexView, StyleguideButtonsComponent, StyleguideView],
+	declarations: [DevView],
 	imports: [
-		ComponentsModule,
 		RouterModule.forChild([
 			{
 				children: devRoutes,
 				component: DevView,
 				path: ""
 			}
-		]),
-		MatDividerModule,
-		MatExpansionModule
+		])
 	]
 })
 export class DevModule {}
