@@ -1,7 +1,11 @@
 import { Component } from "@angular/core";
+import { RouterModule } from "@angular/router";
 
 @Component({
+	standalone: true,
 	styleUrls: ["./index.view.scss"],
-	templateUrl: "./index.view.html"
+	templateUrl: "./index.view.html",
+
+	imports: [RouterModule]
 })
 export class IndexView {}
