@@ -1,4 +1,3 @@
 import * as LOCALE_EN from "./en.json";
-import * as LOCALE_FR from "./fr.json";
 
-export { LOCALE_EN, LOCALE_FR };
+export { LOCALE_EN };
