@@ -6,5 +6,8 @@ const path = require("path");
  */
 module.exports = {
 	...nxPreset,
-	setupFilesAfterEnv: [path.join(__dirname, "./tools/jest/jest-extended.ts")]
+	coverageReporters: ["text"],
+	setupFilesAfterEnv: [path.join(__dirname, "./tools/jest/jest-extended.ts")],
+	// eslint-disable-next-line @cspell/spellchecker -- plugin names
+	watchPlugins: ["jest-watch-typeahead/filename", "jest-watch-typeahead/testname"]
 };
