@@ -1,7 +1,11 @@
 import { JestConfigWithTsJest } from "ts-jest";
 
 export default {
-	coverageDirectory: "../../coverage/libs/common",
+	collectCoverageFrom: [
+		"<rootDir>/src/**/*.ts",
+		"!<rootDir>/src/**/index.ts",
+		"!<rootDir>/src/seeds/**"
+	],
 	displayName: "lib",
 	moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
 	preset: "../../jest.preset.js",
