@@ -1,20 +1,29 @@
 import { Injectable } from "@angular/core";
-import * as qs from "qs/lib/stringify";
-import { EntityId } from "~/app/common/dtos/_lib/entity";
-import { EntityFilter } from "~/app/common/endpoints/_lib";
-import {
-	EntityFindQuery,
-	EntityFindResult
-} from "~/app/common/endpoints/_lib/entity-find.interfaces";
-import { EntityEndpoint } from "~/app/common/endpoints/_lib/entity.endpoint";
+import stringify from "qs/lib/stringify";
+import { Jsonify } from "type-fest";
+import { EntityDto, EntityId } from "~/lib/common/dtos/entity";
+import { DtoToEntity } from "~/lib/common/dtos/entity/entity.types";
+import { EntityFindQuery, EntityFindResult } from "~/lib/common/endpoints/entity-find.interfaces";
+import { EntityEndpoint } from "~/lib/common/endpoints/entity.endpoint";
 
 import { ApiClient } from "../../api.client";
+
+export interface FindAndCountParams {
+	/**
+	 * The uri where the find request should be done
+	 */
+	uri: string;
+}
 
 @Injectable({
 	providedIn: "root"
 })
-export abstract class EntityApiService<T, ToCreate, ToUpdate, Q extends T = T>
-	implements EntityEndpoint<T, ToCreate, ToUpdate, Q>
+export abstract class EntityApiService<
+	T extends DtoToEntity<EntityDto> | Jsonify<EntityDto>,
+	ToCreate,
+	ToUpdate,
+	Q extends EntityFindQuery<T> = EntityFindQuery<T>
+> implements EntityEndpoint<T, ToCreate, ToUpdate, Q>
 {
 	public constructor(protected readonly client: ApiClient) {}
 
@@ -32,8 +41,11 @@ export abstract class EntityApiService<T, ToCreate, ToUpdate, Q extends T = T>
 	 * @param where The filter to apply
 	 * @returns The number of entities found
 	 */
-	public count(where?: EntityFilter<Q>): Promise<number> {
-		return this.findAndCount({ limit: 0, where }).then(({ pagination: { total } }) => total);
+	public count(where?: Q["where"]): Promise<number> {
+		// eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- Subtype constraint
+		return this.findAndCount({ limit: 0, where } as Q).then(
+			({ pagination: { total } }) => total
+		);
 	}
 
 	/**
@@ -42,16 +54,8 @@ export abstract class EntityApiService<T, ToCreate, ToUpdate, Q extends T = T>
 	 * @param query Filter, sort and/or paginate the results
 	 * @returns The results of the request
 	 */
-	public findAndCount(query?: EntityFindQuery<Q>): Promise<EntityFindResult<T>> {
-		let url = this.getEntrypoint();
-		if (query) {
-			const queryString = qs.stringify(query);
-			if (queryString) {
-				url += `?${queryString}`;
-			}
-		}
-
-		return this.client.get(url);
+	public findAndCount(query?: Q): Promise<EntityFindResult<T>> {
+		return this.findAndCountFromParams({ uri: this.getEntrypoint() }, query);
 	}
 
 	/**
@@ -94,5 +98,30 @@ export abstract class EntityApiService<T, ToCreate, ToUpdate, Q extends T = T>
 	 */
 	public delete(id: EntityId): Promise<T> {
 		return this.client.delete(`${this.getEntrypoint()}/${id}`);
+	}
+
+	/**
+	 * Does 'find' request.
+	 * The url can be different, but the filter are the same:
+	 * ex: '/users' and '/group/1/users'
+	 *
+	 * @param params specific parameters
+	 * @param query to request
+	 * @returns the data found and its total
+	 */
+	protected findAndCountFromParams<T2 = T, Q2 extends EntityFindQuery<T2> = EntityFindQuery<T2>>(
+		params: FindAndCountParams,
+		query?: Q2
+	): Promise<EntityFindResult<T2>> {
+		let { uri } = params;
+
+		if (query) {
+			const queryString = stringify(query);
+			if (queryString) {
+				uri += `?${queryString}`;
+			}
+		}
+
+		return this.client.get(uri);
 	}
 }
