@@ -1,2 +1,1 @@
-export * from "./user-api.module";
-export * from "./user-api.service";
+export * from "./user.api.service";
