@@ -1,4 +1,3 @@
 import * as LOCALE_LIB_EN from "./en.json";
-import * as LOCALE_LIB_FR from "./fr.json";
 
-export { LOCALE_LIB_EN, LOCALE_LIB_FR };
+export { LOCALE_LIB_EN };
