@@ -1,4 +1,5 @@
 import { TestBed } from "@angular/core/testing";
+import { TranslateTestingModule } from "ngx-translate-testing";
 
 import { TranslationService } from "./translation.service";
 
@@ -6,7 +7,7 @@ describe("TranslationService", () => {
 	let service: TranslationService;
 
 	beforeEach(() => {
-		TestBed.configureTestingModule({});
+		TestBed.configureTestingModule({ imports: [TranslateTestingModule.withTranslations({})] });
 		service = TestBed.inject(TranslationService);
 	});
 
