@@ -1,4 +1,5 @@
 // https://github.com/ljharb/qs/issues/270#issuecomment-496173839
 declare module "qs/lib/stringify" {
-	export { stringify } from "qs";
+	// eslint-disable-next-line import/no-default-export -- See GH issue
+	export { stringify as default } from "qs";
 }
